bid: divide the code in multiples files
