+++ sources/libraries/brutal/bid/bid.c
@@ -1,6 +1,7 @@
 #include <brutal/alloc.h>
 #include <brutal/base.h>
 #include <brutal/bid/bid.h>
+#include <brutal/bid/bid_interface.h>
 #include <brutal/log.h>
 #include <ctype.h>
 #include "brutal/alloc/base.h"
@@ -21,6 +22,30 @@ static BidResult create_bid(Str data_in)
     scan_init(&result.scanner, data_in);
 
     return OK(BidResult, result);
+}
+
+struct bid_error bid_create_error(enum bid_error_type type, Str error_msg, const struct bid *idl_in)
+{
+    return (struct bid_error){
+        .message = error_msg,
+        .pos = (struct bid_buffer_position){idl_in->scanner.head},
+        .type = type};
+}
+
+struct bid_error bid_create_unexpected_token_error(Str expected_token, const struct bid *idl_in)
+{
+    struct bid_error err_res = bid_create_error(BID_ERROR_TYPE_UNEXPECTED_TOKEN, str_cast("unexpected token"), idl_in);
+    err_res.specific_information.expected_token.name = expected_token;
+    return err_res;
+}
+
+int bid_is_keyword(int chr)
+{
+    if (isalnum(chr))
+    {
+        return true;
+    }
+    return (chr == '_');
 }
 
 static bool skip_comment(struct bid *idl_in)
@@ -52,7 +77,7 @@ static bool skip_comment(struct bid *idl_in)
     return false; // clang warns fix (unreachable code)
 }
 
-static void skip_comment_and_space(struct bid *idl_in)
+void skip_comment_and_space(struct bid *idl_in)
 {
     // loop it until we don't hit a comment or a space
 
@@ -64,319 +89,6 @@ static void skip_comment_and_space(struct bid *idl_in)
             return;
         }
     };
-}
-
-static struct bid_ast_node *create_ast_node(enum bid_ast_node_type type)
-{
-    struct bid_ast_node *node = alloc_malloc(alloc_global(), sizeof(struct bid_ast_node));
-    vec_init(&node->children, alloc_global());
-    node->type = type;
-    return node;
-}
-
-static void destroy_ast_node_recursive(struct bid_ast_node *from)
-{
-
-    if (from != NULL)
-    {
-        for (int i = 0; i < from->children.length; i++)
-        {
-            destroy_ast_node_recursive(from->children.data[i]);
-        }
-        vec_deinit(&from->children);
-        alloc_free(alloc_global(), from);
-    }
-}
-
-void print_ast_node_recursive(struct bid_ast_node *from)
-{
-
-    if (from != NULL)
-    {
-        for (int i = 0; i < from->children.length; i++)
-        {
-            print_ast_node_recursive(from->children.data[i]);
-        }
-
-        log("ast type: {}", from->type);
-    }
-}
-
-int is_keyword(int chr)
-{
-    if (isalnum(chr))
-    {
-        return true;
-    }
-    return (chr == '_');
-}
-
-static struct bid_error bid_create_error(enum bid_error_type type, Str error_msg, const struct bid *idl_in)
-{
-    return (struct bid_error){
-        .message = error_msg,
-        .pos = (struct bid_buffer_position){idl_in->scanner.head},
-        .type = type};
-}
-
-static struct bid_error bid_create_unexpected_token_error(Str expected_token, const struct bid *idl_in)
-{
-    struct bid_error err_res = bid_create_error(BID_ERROR_TYPE_UNEXPECTED_TOKEN, str_cast("unexpected token"), idl_in);
-    err_res.specific_information.expected_token.name = expected_token;
-    return err_res;
-}
-static BidParseResult scan_type_arg(struct bid *idl_in, struct bid_ast_node *type_node)
-{
-    // [type]
-    skip_comment_and_space(idl_in);
-    Str name = scan_skip_until((&idl_in->scanner), is_keyword);
-
-    if (name.len == 0)
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_ALNUM_STR, idl_in));
-    }
-    type_node->ntype.name = name;
-    skip_comment_and_space(idl_in);
-
-    // type[<]
-    if (scan_curr(&idl_in->scanner) == '<')
-    {
-        scan_next(&idl_in->scanner);
-        // type<...>
-        while (scan_curr(&idl_in->scanner) != '>' && !scan_end(&idl_in->scanner))
-        {
-
-            auto ast = create_ast_node(BID_AST_NODE_TYPE);
-            vec_push(&type_node->children, ast);
-
-            // type<[type]>
-            scan_type_arg(idl_in, ast);
-
-            // type<type[,]type>
-            if (scan_curr(&idl_in->scanner) == ',')
-            {
-
-                scan_next(&idl_in->scanner);
-            }
-            // type<type[>]
-            else if (scan_curr(&idl_in->scanner) != '>')
-            {
-
-                return ERR(BidParseResult, bid_create_unexpected_token_error(BID_END_TYPE, idl_in));
-            }
-        }
-
-        if (scan_end(&idl_in->scanner))
-        {
-            return ERR(BidParseResult, bid_create_unexpected_token_error(BID_GREATER_THAN, idl_in));
-        }
-        scan_next(&idl_in->scanner); // skip '>'
-    }
-    return OK(BidParseResult, (MonoState){});
-}
-static BidParseResult scan_method_arg(struct bid *idl_in, struct bid_ast_node *arg_node)
-{
-    // [name] : type
-    skip_comment_and_space(idl_in);
-    Str name = scan_skip_until((&idl_in->scanner), is_keyword);
-
-    if (name.len == 0)
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_ALNUM_STR, idl_in));
-    }
-    arg_node->argument.name = name;
-
-    skip_comment_and_space(idl_in);
-
-    // name [:] type
-    if (scan_curr(&idl_in->scanner) != ':')
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_DOUBLE_DOT, idl_in));
-    }
-    scan_next(&idl_in->scanner);
-    skip_comment_and_space(idl_in);
-
-    // name : [type]
-    auto ast = create_ast_node(BID_AST_NODE_TYPE);
-    vec_push(&arg_node->children, ast);
-
-    TRY(BidParseResult, scan_type_arg(idl_in, ast));
-    return OK(BidParseResult, (MonoState){});
-}
-static BidParseResult scan_method(struct bid *idl_in, struct bid_ast_node *method_node)
-{
-    // method [name](..)->...;
-    skip_comment_and_space(idl_in);
-    Str name = scan_skip_until((&idl_in->scanner), is_keyword);
-
-    if (name.len == 0)
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_ALNUM_STR, idl_in));
-    }
-    method_node->method.name = name;
-    // method name[(]...)->...;
-    skip_comment_and_space(idl_in);
-    if (scan_curr(&idl_in->scanner) != '(')
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_OPENNING_PARENTHESIS, idl_in));
-    }
-
-    // method name([...])->...;
-    scan_next(&idl_in->scanner);
-    while (scan_curr(&idl_in->scanner) != ')' && !scan_end(&idl_in->scanner))
-    {
-        // method name(arg)->...;
-        skip_comment_and_space(idl_in);
-        auto ast = create_ast_node(BID_AST_NODE_TYPE_METHOD_ARGUMENT_TYPE);
-        vec_push(&method_node->children, ast);
-
-        TRY(BidParseResult, scan_method_arg(idl_in, ast));
-
-        skip_comment_and_space(idl_in);
-
-        // method name(arg[,] )->...;
-        if (scan_curr(&idl_in->scanner) == ',')
-        {
-
-            scan_next(&idl_in->scanner);
-            skip_comment_and_space(idl_in);
-        }
-        // method name(arg[)]->...;
-        else if (scan_curr(&idl_in->scanner) != ')')
-        {
-
-            return ERR(BidParseResult, bid_create_unexpected_token_error(BID_END_ARGUMENT, idl_in));
-        }
-    }
-    if (scan_end(&idl_in->scanner))
-    {
-
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_CLOSING_PARENTHESIS, idl_in));
-    }
-    scan_next(&idl_in->scanner);
-    skip_comment_and_space(idl_in);
-
-    // method name()[;]
-    if (scan_curr(&idl_in->scanner) == ';')
-    {
-        scan_next(&idl_in->scanner); // ;
-        return OK(BidParseResult, (MonoState){});
-    }
-
-    // method name() [->] type;
-
-    if (scan_curr(&idl_in->scanner) != '-' || scan_peek(&idl_in->scanner, 1) != '>')
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_ARROW, idl_in));
-    }
-    scan_next(&idl_in->scanner); // skip [-]>
-    scan_next(&idl_in->scanner); // skip -[>]
-    skip_comment_and_space(idl_in);
-
-    auto ast = create_ast_node(BID_AST_NODE_TYPE_METHOD_RETURN_TYPE);
-    vec_push(&method_node->children, ast);
-
-    TRY(BidParseResult, scan_type_arg(idl_in, ast));
-
-    skip_comment_and_space(idl_in);
-
-    // method name() -> type[;]
-    if (scan_curr(&idl_in->scanner) != ';')
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_END_LINE, idl_in));
-    }
-    scan_next(&idl_in->scanner); // ;
-    return OK(BidParseResult, (MonoState){});
-}
-
-static BidParseResult scan_interface_block(struct bid *idl_in, struct bid_ast_node *interface_node)
-{
-    (void)interface_node;
-
-    // interface name [{]}
-    if (scan_curr(&idl_in->scanner) != '{')
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_OPENNING_BRACKETS, idl_in));
-    }
-    scan_next(&idl_in->scanner);
-    skip_comment_and_space(idl_in);
-
-    // interface name  { [...] }
-    while (scan_curr(&idl_in->scanner) != '}' && !scan_end(&idl_in->scanner))
-    {
-
-        skip_comment_and_space(idl_in);
-
-        Str result = scan_skip_until((&idl_in->scanner), is_keyword);
-
-        if (result.len == 0 && scan_curr(&idl_in->scanner) == '}')
-        {
-            break;
-        }
-        else if (result.len == 0 && scan_curr(&idl_in->scanner) != '}')
-        {
-            return ERR(BidParseResult, bid_create_unexpected_token_error(BID_ALNUM_STR, idl_in));
-        }
-        else if (str_eq(result, str_cast("method")))
-        {
-
-            auto ast = create_ast_node(BID_AST_NODE_TYPE_METHOD);
-            vec_push(&interface_node->children, ast);
-
-            TRY(BidParseResult, scan_method(idl_in, ast));
-        }
-        else
-        {
-            return ERR(BidParseResult, bid_create_unexpected_token_error(str_cast("method/struct/enum/typedef"), idl_in));
-        }
-    }
-
-    // interface name  { [}]
-    if (scan_curr(&idl_in->scanner) != '}')
-    {
-
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_CLOSING_BRACKETS, idl_in));
-    }
-    skip_comment_and_space(idl_in);
-
-    return OK(BidParseResult, (MonoState){});
-}
-
-static BidParseResult scan_interface_definition(struct bid *idl_in)
-{
-
-    skip_comment_and_space(idl_in);
-
-    Str result = scan_skip_until((&idl_in->scanner), is_keyword);
-
-    // [interface] name { }
-    if (str_eq(result, str_cast("interface")))
-    {
-        auto ast = create_ast_node(BID_AST_NODE_TYPE_INTERFACE);
-        idl_in->root_ast = ast; // for the moment support 1 interface per file
-
-        skip_comment_and_space(idl_in);
-
-        // interface [name] { }
-        Str interface_name = scan_skip_until((&idl_in->scanner), is_keyword);
-
-        if (interface_name.len == 0)
-        {
-            return ERR(BidParseResult, bid_create_unexpected_token_error(BID_ALNUM_STR, idl_in));
-        }
-
-        ast->interface.name = interface_name;
-
-        skip_comment_and_space(idl_in);
-        TRY(BidParseResult, scan_interface_block(idl_in, ast));
-        return OK(BidParseResult, (MonoState){});
-    }
-    else
-    {
-        return ERR(BidParseResult, bid_create_unexpected_token_error(BID_INTERFACE_STR, idl_in));
-    }
-
-    return OK(BidParseResult, (MonoState){});
 }
 
 static BidParseResult scan_bid(struct bid *idl_in)
