+++ sources/libraries/brutal/bid/bid_convert_c.c
@@ -1,6 +1,7 @@
 #include <brutal/alloc.h>
 #include <brutal/base.h>
 #include <brutal/bid/bid.h>
+#include <brutal/bid/bid_ast.h>
 #include <brutal/bid/bid_convert_c.h>
 #include <brutal/log.h>
 #include "brutal/text/str.h"
